Move loggers into a package
