+++ src/main/java/com/olgaboiar/mint/loggers/ConsoleLogger.java
@@ -1,4 +1,4 @@
-package com.olgaboiar.mint;
+package com.olgaboiar.mint.loggers;
 
 import java.io.IOException;
