+++ src/main/java/com/olgaboiar/mint/loggers/FileLogger.java
@@ -1,4 +1,4 @@
-package com.olgaboiar.mint;
+package com.olgaboiar.mint.loggers;
 
 import java.io.File;
 import java.io.FileOutputStream;
