+++ src/main/java/com/olgaboiar/mint/loggers/ILogger.java
@@ -1,4 +1,4 @@
-package com.olgaboiar.mint;
+package com.olgaboiar.mint.loggers;
 
 import java.io.IOException;
